+++ CHANGELOG.md
@@ -28,7 +28,7 @@
 * Scatter Plot: Added support for DateTime types (#3154)
 * Style: Added support for more line patterns (#3161) _Thanks @MCF_
 * Controls: Assigning `Interaction` can be used instead of `Replace()` for customizing mouse actions (#3150)
-* Controls: Added `IMenu` for simpler context menu customization (#3150)
+* Controls: Added `Menu` with `Add()` and `Clear()` methods to simplify context menu customization (#3150)
 
 ## ScottPlot 5.0.11-beta
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2023-12-29_